shift caching into the topology crate
