+++ common/client-core/src/client/topology_control/nym_api_provider.rs
@@ -3,8 +3,7 @@
 
 use async_trait::async_trait;
 use log::{debug, error, warn};
-use nym_topology::provider_trait::TopologyProvider;
-use nym_topology::NymTopology;
+use nym_topology::{TopologyProvider, NymTopology};
 use nym_validator_client::UserAgent;
 use rand::prelude::SliceRandom;
 use rand::thread_rng;
